Refactor location Controllers
Refactor ubication to controllers for html pages
